pyproject.toml
[tool.poetry]
name = "api-message-sender"
version = "0.1.0"
description = ""
authors = ["lazzarusxd <joaolazarofera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.10.3"}
asyncpg = "^0.30.0"
pytz = "^2024.2"
python-dotenv = "^1.0.1"
aio-pika = "^9.5.3"
aiosmtplib = "^3.0.2"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
